+++ bonfo/msp/structs/__init__.py
@@ -1,0 +1,15 @@
+from .config import CopyProfile, EepromWrite, RcTuning, RxConfig, SelectSetting
+from .status import (
+    AccTrim,
+    ApiVersion,
+    BoardInfo,
+    BuildInfo,
+    FcVariant,
+    FcVersion,
+    Name,
+    RawIMU,
+    SensorAlignment,
+    Status,
+    StatusEx,
+    Uid,
+)