+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.1 (2022-04-XX)
 
 * Added profile management to board class
+* switched to asyncio
 
 ## 0.1.0 (2022-03-29)